requirements.txt
python-dotenv==1.0.0
APScheduler==3.10.4
aiogram==3.3.0
